Move StaticHeader into its own file
