+++ src/engine/strat_engine/backstore/metadata/mod.rs
@@ -2,11 +2,21 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+// Transform a constant in sectors to a constant in bytes
+#[macro_export]
+macro_rules! bytes {
+    ($number:expr) => {
+        $number * devicemapper::SECTOR_SIZE
+    };
+}
+
 mod bda;
 mod mda;
 mod sizes;
+mod static_header;
 
 pub use self::{
-    bda::{device_identifiers, disown_device, BDA},
+    bda::BDA,
     sizes::{BDAExtendedSize, BlockdevSize, MDADataSize},
+    static_header::{device_identifiers, disown_device},
 };
